+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## unreleased [2023-03-04]
+- describe 4 dots on board
+
 ## unreleased [2023-03-02]
 - change snackbar style,duration