+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## unreleased [2023-03-05]
+- change app icon
+
 ## unreleased [2023-03-04]
 - describe 4 dots on board
 - change delete button color