+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## unreleased [2023-03-10]
+- add Crashlytics
+
 ## v1.0.14 [2023-03-05]
 
 ### [2023-03-05]